The Theme has been created
The theme has been created, there are some changes in the main.dart file that enables the app to use the dark or light mode.

The Dark mode is a real black in stead of being a dark gray. We shall work on it next time.

This update is made to help for a short time, it shall be well implemented next time.